refactor: 빌딩 테스트 코드 리팩토링 (#114)
