+++ src/main/java/ku_rum/backend/domain/building/presentation/BuildingSearchController.java
@@ -7,8 +7,7 @@ import ku_rum.backend.domain.building.application.BuildingSearchService;
 import ku_rum.backend.domain.category.dto.request.BuildindgCategoryRequest;
 import ku_rum.backend.domain.category.dto.response.CategoryDetailResponse;
 import ku_rum.backend.domain.user.application.UserService;
-import ku_rum.backend.global.response.BaseResponse;
-import ku_rum.backend.global.response.status.BaseExceptionResponseStatus;
+import ku_rum.backend.global.support.response.BaseResponse;
 import ku_rum.backend.global.security.jwt.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,7 +18,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Optional;
 
-import static ku_rum.backend.global.response.status.BaseExceptionResponseStatus.SUCCESS;
+import static ku_rum.backend.global.support.response.status.BaseExceptionResponseStatus.SUCCESS;
 
 @Slf4j
 @RestController
