+++ src/main/java/ku_rum/backend/global/exception/building/BuildingNotRegisteredException.java
@@ -1,6 +1,6 @@
 package ku_rum.backend.global.exception.building;
 
-import ku_rum.backend.global.response.status.BaseExceptionResponseStatus;
+import ku_rum.backend.global.support.response.status.BaseExceptionResponseStatus;
 import lombok.Getter;
 
 @Getter
